Add Python version of common schema.
In Python, there's a strong pressure to denormalize the tables (and in
particular, use compound keys for joins); that lets us work with the
objects without having to obtain the appropriate autoincrement IDs
from a database or assume that different databases have the same IDs.
But it's becoming clear that might preclude generating the SQL from
the Python definition or vice-versa.